Pass Trusty memory region to OVMF through E820 map
Trusty on ACRN requires a special memory region in the guest for its
secure world (currently 511 GB - 512 GB). In order for this memory
region to be accessible by the guest OS loader, OVMF needs to properly
set up a valid page table mapping for it.

Pass this information as an unusable E820 entry to OVMF. OVMF will mark
this region as reserved memory space and also ensure the EFI virtual
memory space includes this region. This also guarantees the region won't
be used as dynamic memory.