hv: io_req: explicitly interpret REQ_PCICFG ioreq during completion
A REQ_MMIO ioreq has a 64-bit value field, whereas a REQ_PCICFG ioreq
only has a 32-bit value field at the same byte offset (24). This means
when we unconditionally interpret the converted REQ_PCICFG ioreq as a
REQ_MMIO ioreq during completion, its next field, i.e. int32_t bus, gets
interpreted as the upper 32 bits of the 64-bit value field in its
REQ_MMIO form. This is fine for now since we use the little-endian
representation and only support PCI bus 0, but it's only a fluke.

Note that this also affects the conversion of a REQ_MMIO ioreq into a
REQ_PCICFG ioreq in the SOS. The upper 32 bits of the value field is
replaced with the bus field. A PCI config space operation can only be up
to 4 bytes in size.

A REQ_PORTIO VHM ioreq may also be converted into a REQ_PCICFG ioreq,
but it doesn't experience this issue because its value field has the
same size and byte offset as a REQ_PCICFG ioreq.

- For safety and clarity, retrieve the value from a converted REQ_PCICFG
  ioreq explicitly during ioreq completion
- Only do so when it's a read ioreq, since a write ioreq doesn't update
  its own value field
- Complain when encountering a vcpu ioreq with type REQ_PCICFG, since
  only a VHM ioreq can be converted into a REQ_PCICFG ioreq

Signed-off-by: Peter Fang <peter.fang@intel.com>